Infrastructure - Linode Server and Domain Linkage
Used terraform to create a very small nanode that has the "infinito27.com" domain linked to it. Bought the domain on Namecheap, and now using terraform, I can automatically link it after adding the appropriate name servers to Namecheap. Its important to mention that the "main.tf" file depends on some variables present on the file "terraform.tfvars". However, since that file works very similarly to a .env file in programming, its not included in the repo. Also decided to document the commands used to provision the infrastructure for the site in Readme.md