Provisioning - Ansible + Terraform
Set the process of provisioning up so that Terraform just generates the hosts file for Ansible (using a template) and then Ansible can go ahead and SSH into the instance and execute the required commands for configuration. I also had to modify the ansible docker image to now have an entrypoint that simply executes the "provision.yaml" playbook. Decided to include a docker-compose file, so that the command to run is a bit shorter (just "docker compose up" instead of 'docker run ansible bash -c "blah blah blah"')