Bugfix - Correct Transfer of Frontend Files
The files for the frontend weren't being copied correctly to the website's server. Aparently the only thing that I had to do was remove the "remote_src" option (since that meant that it was going to look for a "opt/frontend" folder inside of the server, instead of in the Ansible container where it was. I also changed the root of the infinito27.com configuration file, so that it serves files from a directory inside of "var/www/infinito27" called "frontend". Now you can go to "infinito 27" and you will find properly rendered page. Its all coming together. Next step, clone the repo into the remote server, so that it can start using CI/CD in the future.