Provisioning - Github Repo
Made it so that the contents of the website come from the github repo itself, not from a copy of the files found in the ansible container itself. This is made to be more compatible with CI/CD in the future. Also deleted the volume mount for the frontend from the Ansible container, since its no longer going to copy those files into the remote host.