Website - Vue + Vite Initial Files
Decided to use a combination of Vite and Vue for the creation of the site (Mostly to learn), so I used the vite launcher to create a couple of initial files. The site has been tested and I even included some actual unit tests using Jest. I implemented those to add the CI step to the github actions pipeline. Thinking of using Dagger soon so that I can create a platform agnostic CI/CD pipeline thats also testable locally.