Separate Frontend and Backend Folders
Decided that I would probably add a backend to this site in the future, so I made the necessary changes in order to create a folder for both frontend and backend (changed the place where NGINX serves the files, as well as changed the place where the files are tested in the CI and CD steps). Hoping that the CI and CD steps are still both working.