Mounted Local Hosts and Ansible.cfg
Included the "hosts" and "ansible.cfg" files from the local directory to prevent the Ansible container from provisioning the host found in the "hosts" file used when building the ansible container. Had the problem where I basically provisioned "infinito27.com" with the configuration file of "eddysanoli.com". To solve this, I decided to add another mount to the docker-compose.yaml thats supposed to run when you want to provision.