Finished Tutorial 4
Finally finished the fourth tutorial. He covered a lot of things, from event modifiers, to emitters, to slots and other things. Its very probable that I will never use all of this stuff, but this is simply to be aware of the tools that I have available in order to know what I can do to build cool stuff. Also, since my mantra is to create a component for each lesson, I put all the logic that would be inside of "App.vue" inside of lesson, and then putting the sublogic of the component that was created in the lesson (a modal) inside its own component. Now we can simply detach the lesson 4 component and everything will start fresh.