Skipped Tutorial 9
The 9th tutorial was a simple tutorial that showed how you fetch and display data. Its not something thats super inherent to Vue, so I decided to skip it. In the future I can go back to it and check what it covered, but from what I skimmed, it wasnt too important. What I did do, is install the json-server package that the tutorial guy uses to mock an API. I dont think I will use it, but I will leave it here just in case.