Pocketbase Executable
Decided to include pocketbase as the backend for the site. Here I can put things like data, images, and other things. Seems like the installation is really simple, just download the binary and do "./pocketbase". Its really that simple. Pretty cool. However, I wont include the binaries themselves, as they are quite small, but in the future could get kind of bloated. For now Im just pushing a gitignore for the pocketbase executables.