Installed Tailwind
Went through a video tutorial (https://www.youtube.com/watch?v=TsQVdEekvtE&ab_channel=TheNetNinja) and followed the instructions to setting up Tailwind. This installation uses a lot of stuff that I didnt know you could do before (like using an @ operator in CSS sheets or using npx to run a tailwind script to generate the default configuration files for it). In the configuration file for Tailwind I included the fonts that are going to be used, as well as some of the colors that I intend to use for the design.